Added getTransactionEncodedData util function
