+++ src/redux/sagas/actions/initiateSafeTransaction.ts
@@ -3,7 +3,6 @@ import { call, fork, put, takeEvery } from 'redux-saga/effects';
 
 import { ContextModule, getContext } from '~context';
 import {
-  SafeTransactionType,
   CreateSafeTransactionMutation,
   CreateSafeTransactionDocument,
   CreateSafeTransactionMutationVariables,
@@ -15,7 +14,6 @@ import { ActionTypes } from '~redux/actionTypes';
 import { Action, AllActions } from '~redux/types';
 import { putError, takeFrom } from '~utils/saga/effects';
 import { fill, omit } from '~utils/lodash';
-import { isDev } from '~constants';
 
 import { transactionReady } from '../../actionCreators';
 import {
@@ -24,13 +22,8 @@ import {
   getTxChannel,
 } from '../transactions';
 import {
-  getRawTransactionData,
-  getTransferNFTData,
-  getTransferFundsData,
-  getContractInteractionData,
-  getZodiacModule,
   getHomeBridgeByChain,
-  ZODIAC_BRIDGE_MODULE_ADDRESS,
+  getTransactionEncodedData,
 } from '../utils/safeHelpers';
 import { uploadAnnotation } from '../utils';
 
@@ -52,69 +45,14 @@ function* initiateSafeTransactionAction({
     txChannel = yield call(getTxChannel, metaId);
     const apolloClient = getContext(ContextModule.ApolloClient);
 
-    const zodiacBridgeModuleAddress = isDev
-      ? ZODIAC_BRIDGE_MODULE_ADDRESS
-      : safe.moduleContractAddress;
-
-    if (!zodiacBridgeModuleAddress) {
-      throw new Error(
-        `Please provide a ZODIAC_BRIDGE_MODULE_ADDRESS. If running local, please add key-pair to your .env file.`,
-      );
-    }
     const homeBridge = getHomeBridgeByChain(safe.chainId);
-    const zodiacBridgeModule = getZodiacModule(zodiacBridgeModuleAddress, safe);
 
-    const transactionData: string[] = [];
-    /*
-     * Calls HomeBridge for each Tx, with the Colony as the sender.
-     * Loop necessary as yield cannot be called inside of an array iterator (like forEach).
-     */
-    /* eslint-disable-next-line no-restricted-syntax */
-    for (const transaction of transactions) {
-      let txDataToBeSentToZodiacModule = '';
-      switch (transaction.transactionType) {
-        case SafeTransactionType.RawTransaction:
-          txDataToBeSentToZodiacModule = getRawTransactionData(
-            zodiacBridgeModule,
-            transaction,
-          );
-          break;
-        case SafeTransactionType.TransferNft:
-          txDataToBeSentToZodiacModule = getTransferNFTData(
-            zodiacBridgeModule,
-            safe,
-            transaction,
-          );
-          break;
-        case SafeTransactionType.TransferFunds:
-          txDataToBeSentToZodiacModule = yield getTransferFundsData(
-            zodiacBridgeModule,
-            safe,
-            transaction,
-            network,
-          );
-          break;
-        case SafeTransactionType.ContractInteraction:
-          txDataToBeSentToZodiacModule = yield getContractInteractionData(
-            zodiacBridgeModule,
-            safe,
-            transaction,
-          );
-          break;
-        default:
-          throw new Error(
-            `Unknown transaction type: ${transaction.transactionType}`,
-          );
-      }
-
-      /* eslint-disable-next-line max-len */
-      const txDataToBeSentToAMB = yield homeBridge.interface.encodeFunctionData(
-        'requireToPassMessage',
-        [zodiacBridgeModule.address, txDataToBeSentToZodiacModule, 1000000],
-      );
-
-      transactionData.push(txDataToBeSentToAMB);
-    }
+    const transactionData: string[] = yield getTransactionEncodedData(
+      transactions,
+      safe,
+      network,
+      homeBridge,
+    );
 
     const batchKey = 'initiateSafeTransaction';
